. changed SteteId into size_t type... which looses the  kNoStateId state .   still set it up to UINT_MAX . . applied ATTRIBUTE_UNUSED macro when needed.
# Conflicts:
#	src/include/fst/arc-map.h
#	src/include/fst/arc.h
#	src/include/fst/closure.h
#	src/include/fst/compat.h
#	src/include/fst/compose-filter.h
#	src/include/fst/concat.h
#	src/include/fst/connect.h
#	src/include/fst/determinize.h
#	src/include/fst/dfs-visit.h
#	src/include/fst/equivalent.h
#	src/include/fst/filter-state.h
#	src/include/fst/fst.h
#	src/include/fst/randgen.h
#	src/include/fst/relabel.h
#	src/include/fst/replace.h
#	src/include/fst/rmepsilon.h
#	src/include/fst/shortest-distance.h
#	src/include/fst/shortest-path.h
#	src/include/fst/state-map.h
#	src/include/fst/symbol-table.h
#	src/include/fst/union.h
#	src/include/fst/vector-fst.h
#	src/include/fst/verify.h

+++ src/include/fst/connect.h
@@ -146,7 +146,7 @@ class SccVisitor {
   void FinishVisit() {
     // Numbers SCCs in topological order when acyclic.
     if (scc_) {
-      for (StateId s = 0; s < scc_->size(); ++s) {
+      for (StateId s = 0; (size_t)s < scc_->size(); ++s) {
         (*scc_)[s] = nscc_ - 1 - (*scc_)[s];
       }
     }
@@ -201,7 +201,7 @@ inline void SccVisitor<Arc>::InitVisit(const Fst<Arc> &fst) {
 template <class Arc>
 inline bool SccVisitor<Arc>::InitState(StateId s, StateId root) {
   scc_stack_->push_back(s);
-  while (dfnumber_->size() <= s) {
+  while (dfnumber_->size() <= (size_t)s) {
     if (scc_) scc_->push_back(-1);
     if (access_) access_->push_back(false);
     coaccess_->push_back(false);
